Refactor theme import paths to use unified module export
