+++ frontend/src/components/JoinContainerSeparator/JoinContainerSeparator.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import Stack from '@ui/Stack';
 import Typography from '@ui/Typography';
-import useCustomTheme from '@Context/Theme/CustomTheme';
+import useCustomTheme from '@Context/Theme';
 import Separator from '../Separator';
 
 /**
