+++ frontend/src/components/LandingPageWelcome/LandingPageWelcome.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import Box from '@ui/Box';
 import Typography from '@ui/Typography';
-import useCustomTheme from '@Context/Theme/CustomTheme';
+import useCustomTheme from '@Context/Theme';
 
 /**
  * LandingPageWelcome Component
